+++ local/php_interface/agents.php
@@ -26,7 +26,7 @@ function AutoPeriodsCreate()
                 ),
                 'filter' => array(
                     'UF_GROUP_ID' => $arGroup['ID'],
-                    //'UF_GROUP_ID' => 29,
+                    //'UF_GROUP_ID' => 3,
                     'UF_AUTO_RENEWAL' => 86,
                     'UF_DEFAULT' => 1
                 )
@@ -52,27 +52,28 @@ function AutoPeriodsCreate()
                 )
             );
 
-            while ($arLastPeriod = $rsPeriod->fetch()) {
-			    /*
-                    dump('arGroup: '.$arGroup['ID']);
-                    dump('arTemplatePeriod: '.$arTemplatePeriod['ID']);
-                    dump('arLastPeriod: '.$arLastPeriod['ID']);
-                    dump('arLastPeriod: '.$arLastPeriod['UF_STATUS']);
-                    dump('arLastPeriod: '.$arLastPeriod['UF_DEFAULT']);
-                    dump('---------------');
-			    */
+            $arLastPeriod = $rsPeriod->fetch();
+            //while ($arLastPeriod = $rsPeriod->fetch()) {
 
-                if ($arLastPeriod['UF_STATUS'] != 'Y' && $arLastPeriod['UF_DEFAULT'] != '1') {
+                /*dump('arGroup: '.$arGroup['ID']);
+                dump('arTemplatePeriod: '.$arTemplatePeriod['ID']);
+                dump('arLastPeriod: '.$arLastPeriod['ID']);
+                dump('arLastPeriod: '.$arLastPeriod['UF_STATUS']);
+                dump('arLastPeriod: '.$arLastPeriod['UF_DEFAULT']);
+                dump('---------------');*/
+
+                //if ($arLastPeriod['UF_STATUS'] != 'Y' && $arLastPeriod['UF_DEFAULT'] != '1') {
+                if ($arLastPeriod['UF_STATUS'] != 'Y') {
 
                     //Перебераем все периоды
                     $endDate = new DateTime($arLastPeriod['UF_DATE_END']);
                     $endDate = strtotime($endDate->format('Y-m-d'));
 
                     $currentDate = new DateTime(date('d-m-Y'));
-                    $currentDate = $currentDate->modify('+2 days');
+                    $currentDate = $currentDate->modify('+15 days');
                     $currentDate = strtotime($currentDate->format('Y-m-d'));
 
-                    //Если дата завершения меньше текущей даты+2дня
+                    //Если дата завершения меньше текущей даты+15дней
                     if ($endDate == $currentDate) {
 
                         $nextDateStart = new DateTime($arLastPeriod['UF_DATE_START']);
@@ -106,7 +107,6 @@ function AutoPeriodsCreate()
                             $nextDateEnd = $nextDateEnd->format('d.m.Y');
                         }
 
-
                         //Формируем поля для нового периода
                         $arPeriodFields = array(
                             'UF_DATE_START' => $nextDateStart->format('d.m.Y'),
@@ -134,11 +134,9 @@ function AutoPeriodsCreate()
                             $totalMustPay = 0;
                             $PERIOD_ID = $result->getId();
 
-                            /*
-                            dump($arGroup['ID']);
+                            /*dump($arGroup['ID']);
                             dump($PERIOD_ID);
-                            dump('----');
-                            */
+                            dump('----');*/
 
                             $arPlatforms = SchedulePayments::getListValues(SchedulePayments::PLATFORM_ID);
                             $arTargets = SchedulePayments::getListValues(SchedulePayments::TARGET_ID);
@@ -170,7 +168,7 @@ function AutoPeriodsCreate()
                         }
                     }
                 }
-            }
+            //}
         }
     }
     return 'AutoPeriodsCreate();';