+++ CHANGELOG.md
@@ -20,6 +20,13 @@ The format is based on [Keep a Changelog][keepachangelog] and this project adher
 > - Каждая запись **должна** завершаться указанием ссылки на того, кто внёс те или иные изменения
 
 
+## v0.0.14 - 2020-05-19
+
+### Fixed
+
+- Fix background-color for duplicate period
+- Fix period auto create from template
+
 ## v0.0.13 - 2020-05-18
 
 ### Fixed